+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # HOF ChangeLog
 
+## 2016-03-03, Version 3.2.0 (Stable), @easternbloc
+
+### Notable changes
+
+* **HOF-Middleware**: Added support for cookie checking on all routes other than /cookies-required
+
 ## 2015-11-25, Version 1.1.0 (Stable), @Joechapman
 
 ### Notable changes